refactored endpoints, added functionality for log out but will be used later, and removed cascade delete

+++ src/main/java/com/gaments/gaments/service/PostServiceImpl.java
@@ -10,6 +10,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpSession;
+import javax.transaction.Transactional;
 import java.util.List;
 
 @Service
@@ -25,7 +27,7 @@ public class PostServiceImpl implements PostService {
     AuthenticationImpl authenticationImpl;
 
     @Override
-    public Post createPost(Post newPost){
+    public Post createPost(Post newPost) {
         Authentication auth = authenticationImpl.getAuthentication();
         User user = userRepository.findByUsername(auth.getName());
         newPost.setUser(user);
@@ -33,26 +35,33 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
-    public List<Post> listAllPosts(){
+    public List<Post> listAllPosts() {
         return postRepository.findAll();
     }
 
     @Override
-    public List<Post> listLoggedPosts(){
+    public List<Post> listLoggedPosts() {
         Authentication auth = authenticationImpl.getAuthentication();
         User user = userRepository.findByUsername(auth.getName());
         return postRepository.findPostsByUser(user);
     }
 
     @Override
-    public List<Post> listUsersPost(String username){
+    public List<Post> listUsersPost(String username) {
         User user = userRepository.findByUsername(username);
         return postRepository.findPostsByUser(user);
     }
 
     @Override
-    public HttpStatus deleteById(Long postId){
-        postRepository.deleteById(postId);
-        return HttpStatus.OK;
+    @Transactional
+    public HttpStatus deleteById(Long postId) {
+        Authentication auth = authenticationImpl.getAuthentication();
+        Long userId = userRepository.findByUsername(auth.getName()).getId();
+        Long postUserId = postRepository.findPostsById(postId).getUser().getId();
+        if (userId.equals(postUserId)) {
+            postRepository.deleteById(postId);
+            return HttpStatus.OK;
+        }
+        return HttpStatus.FORBIDDEN;
     }
 }
